requirements.txt
python-pptx
requests
beautifulsoup4
browser-cookie3
Pillow
lxml
